+++ packages/terra-core-docs/src/terra-dev-site/doc/status-view/example/Custom.jsx
@@ -15,6 +15,7 @@ const StatusViewButtons = [
 const Custom = () => (
   <StatusView
     message="Custom Message"
+    aria-live="polite" 
     customGlyph={<IconDiamond />}
     title="Custom Title"
     buttonAttrs={StatusViewButtons}